refactor: add collector api parameter types and rename to model.ts
