+++ apps/web/src/services/asset-inventory/collector/collector-create/modules/CreateCollectorStep1.vue
@@ -62,10 +62,10 @@ import { gray } from '@/styles/colors';
 import { BACKGROUND_COLOR } from '@/styles/colorsets';
 
 import Step1SearchFilter from '@/services/asset-inventory/collector/collector-create/modules/Step1SearchFilter.vue';
+import type { RepositoryPluginModel } from '@/services/asset-inventory/collector/model';
 import { useCollectorFormStore } from '@/services/asset-inventory/collector/shared/collector-forms/collector-form-store';
 import CollectPluginContents
     from '@/services/asset-inventory/collector/shared/CollectorPluginContents.vue';
-import type { RepositoryPluginModel } from '@/services/asset-inventory/collector/type';
 
 const emit = defineEmits([
     'update:currentStep',
